insert deprecated feature ids directly into migration.py
