+++ scripts/migration_assistant/migration_template.jinja
@@ -81,8 +81,15 @@ def migrate(input_file, output_file, collection_id, dataset_id):
     # ...
 
     # AUTOMATED, DO NOT CHANGE -- IF GENCODE UPDATED, DEPRECATED FEATURE FILTERING ALGORITHM WILL GO HERE.
-    {% if gencode_migrate %}
-    dataset = utils.remove_deprecated_features(dataset)
+    {% if deprecated_feature_ids %}
+    # fmt: off
+    deprecated_features_ids = [
+        {% for feature in deprecated_feature_ids %}
+        "{{ feature }}",
+        {% endfor %}
+    ]
+    # fmt: on
+    dataset = utils.remove_deprecated_features(dataset, deprecated_features_ids)
     {% else %}
     # No Changes
     {% endif %}
